fix: åtgärda CodeRabbit-kommentarer (licensrubrik, long-cast, dubbelklick) och byt länktext i hjälpsidor
- Lägg till licensrubrik i SearchExportPluginTest.java
- Ändra exportSearchDialogHitCount från int till long i ClientMessages och ExportSearchDialog
- Lägg till exportInProgress-guard i onStartExport för att förhindra dubbelklick
- Byt länktext i Help.md och Help_sv_SE.md till "Konfigurera fält för export av sökresultat"

+++ roda-ui/roda-wui/src/main/java/config/i18n/client/ClientMessages.java
@@ -1994,7 +1994,7 @@ public interface ClientMessages extends Messages {
   String exportSearchDialogTitle();
 
   @DefaultMessage("Your search returned {0} items. All will be exported as CSV.")
-  String exportSearchDialogHitCount(int count);
+  String exportSearchDialogHitCount(long count);
 
   @DefaultMessage("Select fields to include:")
   String exportSearchDialogFieldsLabel();
